# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Millicall PBX",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/workflows": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_api_workflows_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowListResponse"
                  },
                  "type": "array",
                  "title": "Response List Workflows Api Workflows Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_api_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/workflows/node-types": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Node Types",
        "operationId": "get_node_types_api_workflows_node_types_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^workflow$",
              "default": "workflow",
              "title": "Workflow Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/generate": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Generate Workflow",
        "description": "Use Gemini to generate a workflow definition from natural language.",
        "operationId": "generate_workflow_api_workflows_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_api_workflows__workflow_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "workflows"
        ],
        "summary": "Update Workflow",
        "operationId": "update_workflow_api_workflows__workflow_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workflows"
        ],
        "summary": "Delete Workflow",
        "operationId": "delete_workflow_api_workflows__workflow_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ondemandcall/{call_id}": {
      "get": {
        "tags": [
          "ondemand-trigger"
        ],
        "summary": "Trigger Ondemand Call",
        "description": "Trigger a call to the registered phone number. No authentication required.",
        "operationId": "trigger_ondemand_call_ondemandcall__call_id__get",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Call Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type",
            "default": "workflow"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "GenerateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodePosition": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "NodePosition"
      },
...
      "TTSConfig": {
        "properties": {
          "tts_provider": {
            "type": "string",
            "title": "Tts Provider",
            "default": "google"
          },
          "google_tts_voice": {
            "type": "string",
            "title": "Google Tts Voice",
            "default": "ja-JP-Chirp3-HD-Aoede"
          },
          "coefont_voice_id": {
            "type": "string",
            "title": "Coefont Voice Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "TTSConfig"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "number": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^\\d+$",
            "title": "Number"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "workflow_type": {
            "type": "string",
            "pattern": "^workflow$",
            "title": "Workflow Type",
            "default": "workflow"
          },
          "definition": {
            "$ref": "#/components/schemas/WorkflowDefinition-Input",
            "default": {
              "nodes": [],
              "edges": []
            }
          },
          "default_tts_config": {
            "$ref": "#/components/schemas/TTSConfig",
            "default": {
              "tts_provider": "google",
              "google_tts_voice": "ja-JP-Chirp3-HD-Aoede",
              "coefont_voice_id": ""
            }
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "number"
        ],
        "title": "WorkflowCreate"
      },
      "WorkflowDefinition-Input": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNode"
            },
            "type": "array",
            "title": "Nodes",
            "default": []
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEdge"
            },
            "type": "array",
            "title": "Edges",
            "default": []
          }
        },
        "type": "object",
        "title": "WorkflowDefinition"
      },
      "WorkflowDefinition-Output": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNode"
            },
            "type": "array",
            "title": "Nodes",
            "default": []
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEdge"
            },
            "type": "array",
            "title": "Edges",
            "default": []
          }
        },
        "type": "object",
        "title": "WorkflowDefinition"
      },
      "WorkflowEdge": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "sourceHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcehandle"
          },
          "targetHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targethandle"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "target"
        ],
        "title": "WorkflowEdge"
      },
      "WorkflowListResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "extension_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Id"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "node_count": {
            "type": "integer",
            "title": "Node Count"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "number",
          "description",
          "extension_id",
          "workflow_type",
          "enabled",
          "node_count",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowListResponse"
      },
      "WorkflowNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "position": {
            "$ref": "#/components/schemas/NodePosition"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "default": {}
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "position"
        ],
        "title": "WorkflowNode"
      },
      "WorkflowResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "extension_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Id"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "definition": {
            "$ref": "#/components/schemas/WorkflowDefinition-Output"
          },
          "default_tts_config": {
            "$ref": "#/components/schemas/TTSConfig"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "number",
          "description",
          "extension_id",
          "workflow_type",
          "definition",
          "default_tts_config",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowResponse"
      },
      "WorkflowUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "workflow_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Type"
          },
          "definition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowDefinition-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "default_tts_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TTSConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "WorkflowUpdate"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}